Revert "Anil/multistore"
