+++ baseapp/options.go
@@ -133,6 +133,5 @@ func (app *BaseApp) SetStoreLoader(loader StoreLoader) {
 	if app.sealed {
 		panic("SetStoreLoader() on sealed BaseApp")
 	}
-
 	app.storeLoader = loader
 }
